// src/agent/model/providers/copilot/responses/openai-responses-api-types.ts
export type OpenAIResponsesInput = Array<OpenAIResponsesInputItem>

export type OpenAIResponsesInputItem =
  | OpenAIResponsesSystemMessage
  | OpenAIResponsesUserMessage
  | OpenAIResponsesAssistantMessage
  | OpenAIResponsesFunctionCall
  | OpenAIResponsesFunctionCallOutput
  | OpenAIResponsesComputerCall
  | OpenAIResponsesLocalShellCall
  | OpenAIResponsesLocalShellCallOutput
  | OpenAIResponsesReasoning
  | OpenAIResponsesItemReference
  | OpenAIResponsesMcpApprovalResponse

export type OpenAIResponsesIncludeValue =
  | 'web_search_call.action.sources'
  | 'code_interpreter_call.outputs'
  | 'computer_call_output.output.image_url'
  | 'file_search_call.results'
  | 'message.input_image.image_url'
  | 'message.output_text.logprobs'
  | 'reasoning.encrypted_content'

export type OpenAIResponsesIncludeOptions = Array<OpenAIResponsesIncludeValue> | undefined | null

export type OpenAIResponsesSystemMessage = {
  role: 'system' | 'developer'
  content: string
}

export type OpenAIResponsesUserMessage = {
  role: 'user'
  content: Array<
    | { type: 'input_text'; text: string }
    | { type: 'input_image'; image_url: string; detail?: string }
    | { type: 'input_image'; file_id: string; detail?: string }
    | { type: 'input_file'; file_url: string }
    | { type: 'input_file'; filename: string; file_data: string }
    | { type: 'input_file'; file_id: string }
  >
}

export type OpenAIResponsesAssistantMessage = {
  role: 'assistant'
  content: Array<{ type: 'output_text'; text: string }>
  id?: string
}

export type OpenAIResponsesFunctionCall = {
  type: 'function_call'
  call_id: string
  name: string
  arguments: string
  id?: string
}

export type OpenAIResponsesFunctionCallOutput = {
  type: 'function_call_output'
  call_id: string
  output: string
}

export type OpenAIResponsesComputerCall = {
  type: 'computer_call'
  id: string
  status?: string
}

export type OpenAIResponsesLocalShellCall = {
  type: 'local_shell_call'
  id: string
  call_id: string
  action: {
    type: 'exec'
    command: Array<string>
    timeout_ms?: number
    user?: string
    working_directory?: string
    env?: Record<string, string>
  }
}

export type OpenAIResponsesLocalShellCallOutput = {
  type: 'local_shell_call_output'
  call_id: string
  output: string
}

export type OpenAIResponsesItemReference = {
  type: 'item_reference'
  id: string
}

export type OpenAIResponsesMcpApprovalResponse = {
  type: 'mcp_approval_response'
  approval_request_id: string
  approve: boolean
}
// ...
export type OpenAIResponsesReasoning = {
  type: 'reasoning'
  id: string
  encrypted_content?: string | null
  summary: Array<{
    type: 'summary_text'
    text: string
  }>
}